refactor: Correct typo and Update import paths for ReplaceLoginPageModal
