+++ apps/jurumarble/src/components/ReplaceLoginPageModal/index.ts
@@ -1,0 +1,1 @@
+export { default as ReplaceLoginPageModal } from './ReplaceLoginPageModal';
